fix purchase book flow, add decreasing mechanism for quantity

+++ infrastructure/src/main/java/com/sokima/reactive/grpc/bookstore/infrastructure/adapter/persistent/transformer/BookEntityTransformer.java
@@ -65,7 +65,7 @@ public class BookEntityTransformer {
         return entity;
     }
 
-    public List<BookEntity> generateBookEntities(final BookAggregationEntity bookAggregation, final Integer count) {
+    public List<BookEntity> generateBookEntities(final BookAggregationEntity bookAggregation, final Long count) {
         log.trace("Transforming to book entities: {} with count: {}", bookAggregation, count);
         return Stream.generate(IsbnGenerator::generateIsbn)
                 .limit(count)
